Move ValidationTestCase to new testvalidation pkg
