+++ internal/pkg/service/buffer/definition/branch_test.go
@@ -1,6 +1,7 @@
 package definition
 
 import (
+	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/storage/test/testvalidation"
 	"testing"
 
 	"github.com/keboola/keboola-as-code/internal/pkg/service/buffer/definition/key"
@@ -10,7 +11,7 @@ func TestBranch_Validation(t *testing.T) {
 	t.Parallel()
 
 	// Test cases
-	cases := ValidationTestCases[Branch]{
+	cases := testvalidation.TestCases[Branch]{
 		{
 			Name: "empty",
 			Value: Branch{
